extract ChatMessage, WrappedChat
